requirements.txt
Flask
requests
opencv-python-headless
easyocr
gunicorn
werkzeug
python-dotenv
